Move metrics tracker from plugin reporting to shared utilities
The metrics tracker implementation previously lived in src/plugin/src/reporting/metrics-tracker.js even though it only depended on shared helpers. Relocating it to src/shared/utils/metrics.js makes the generic collector available to other packages while keeping the plugin API intact via the shared barrel exports and updated tests.
