+++ src/plugin/src/reporting/index.js
@@ -1,1 +1,1 @@
-export * from "./metrics-tracker.js";
+export * from "../../../shared/metrics.js";
